reference: move throw error above catching it
When reading this documention it makes much more sense to first learn
about the different ways to throw an error (as command for a job and in
the process as end event). These concepts are also immediately used in
the catching the error section, so the reader needs to be aware of it
before.