reference: describe uncaught error incident resolution
Uncaught error incidents attached to service tasks can now be resolved,
which makes the job activatable again. To clarify this it also makes
sense to mention that jobs are 'disabled' when an error is thrown for
them.

The job is removed when the service task is terminated, which only
happens when the error is caught.