reference: improve defining the error
It could be unclear for readers what must be defined. By making it clear
that the error must be defined in BPMN and error events refer to these
errors, I think it's more understandible.

Note that modeller uses this same idea (in the details of an error event
you need to choose which error the event refers to).