reference: replace handle error with catch and deal with
Handling is a strange word. The BPMN spec generally uses the catch an
error terminolgy, which we also already use. I've transitioned to
catching the error in all cases where it is explicitly about the case to
catch the error with an event. When we talk about the error in the
domain and potentially compensation for it, I've moved to dealing with
the error. The only place where handle sadly still arives is in the
incident type `unhandled error event` when the error was not caught.
Since this is something users might search for, I've kept this wording
but only when specifically talking about this incident.

This frees up the usage of handle for other things, like handling a job
in the job worker (which we sometimes also call processing a job),
zeebe's request handling, etc.

From now on, let's talk about dealing with errors and catching them in
an event.