Release 1.1 (#310)
* Describe uncaught error incident handling (#282)

* reference: reformat error events

By splitting the lines at the natural end of each sentence (i.e. at the
dot `.`), the diffs become much smaller to look at during review.

This is not part of our styleguide, but I still apply it, because I
believe it makes sense for the upcoming changes.

* reference: move throw error above catching it

When reading this documention it makes much more sense to first learn
about the different ways to throw an error (as command for a job and in
the process as end event). These concepts are also immediately used in
the catching the error section, so the reader needs to be aware of it
before.

* reference: describe uncaught error incident resolution

Uncaught error incidents attached to service tasks can now be resolved,
which makes the job activatable again. To clarify this it also makes
sense to mention that jobs are 'disabled' when an error is thrown for
them.

The job is removed when the service task is terminated, which only
happens when the error is caught.

* reference: fix typos

* reference: improve grammar

* reference: improve error events intro

By explaining the concept of business errors in more detail as well as
explaining how errors in bpmn must be thrown and caught, I believe this
intro has become more understandible for new readers. This idea was
suggested in the review.

* reference: improve defining the error

It could be unclear for readers what must be defined. By making it clear
that the error must be defined in BPMN and error events refer to these
errors, I think it's more understandible.

Note that modeller uses this same idea (in the details of an error event
you need to choose which error the event refers to).

* reference: improve throwing the error

* reference: improve catching the error

I thought it made sense to describe the error catching and propagation
logic in more detail.

Secondly, the error handling needs to be introduced, otherwise the topic
of unhandled errors don't make any sense.

* reference: improve unhandled errors

* reference: apply review suggestions

Co-authored-by: Peter Ihme <pihme@users.noreply.github.com>

* reference: replace handle error with catch and deal with

Handling is a strange word. The BPMN spec generally uses the catch an
error terminolgy, which we also already use. I've transitioned to
catching the error in all cases where it is explicitly about the case to
catch the error with an event. When we talk about the error in the
domain and potentially compensation for it, I've moved to dealing with
the error. The only place where handle sadly still arives is in the
incident type `unhandled error event` when the error was not caught.
Since this is something users might search for, I've kept this wording
but only when specifically talking about this incident.

This frees up the usage of handle for other things, like handling a job
in the job worker (which we sometimes also call processing a job),
zeebe's request handling, etc.

From now on, let's talk about dealing with errors and catching them in
an event.

* reference: add ignore error when resolving incident

It should also be clear what happens to the error that was not caught
when the incident (`unhandled error event`) attached to a task is
resolved.

Co-authored-by: Peter Ihme <pihme@users.noreply.github.com>

* feat(reference): describe how to use the new tasks (#298)

* add new pages in the BPMN reference to read more about business rule tasks, script tasks, and send tasks

Co-authored-by: Nico Korthout <korthout@users.noreply.github.com>

* feat(reference): update bpmn coverage (#301)

* mark send tasks, business rule tasks, and script tasks as implemented

Co-authored-by: Nico Korthout <korthout@users.noreply.github.com>
Co-authored-by: Peter Ihme <pihme@users.noreply.github.com>
Co-authored-by: Philipp Ossler <philipp.ossler@gmail.com>