Switch orange to green on BPMN coverage
The orange highlighting currently used in the BPMN coverage map
feels like a warning color that one would expect to be used for the unsupported elements.
In addition the color is not a Camunda brand color.
The replacement color is from the official Icemunda color scale
and fits nicely to the hover color.