Describe assignment definitions for user tasks (#462)
* docs(reference): add assignments to user tasks

Adds a description of how assignments can be defined for user tasks.
That either one or both of (assignee and candidate groups) can be
specified. How they can be specified in xml (with example) and what
their values can be (static/expression).

* docs(reference): fix typo

Co-authored-by: Philipp Ossler <philipp.ossler@gmail.com>

* docs(reference): use list as candidate groups example

It should be clear from the example that the candidateGroups as static value can be a comma-separated list of values.

Co-authored-by: Peter Ihme <pihme@users.noreply.github.com>

* docs(reference): add tasklist assignee assumptions

In order for Tasklist to assign an actual Tasklist user, the assignee's
value must be a known user id. This id is dependent on the auth method
used to login to Tasklist.

* core(docs): fix link to tasklist

Co-authored-by: Philipp Ossler <philipp.ossler@gmail.com>
Co-authored-by: Peter Ihme <pihme@users.noreply.github.com>
Co-authored-by: Sebastian Menski <sebastian.menski@camunda.com>