1.4.0 - Zeebe process test (#507)
* Move deprecated zeebe-test to new page

* Add reference to alternative for deprecated zeebe-test

* Add zeebe-process-test to docs

* technical writing review of testing and zeebe test

* Removed zeebe-test page

Zeebe-test was deprecated in 1.3.0 and marked for removal in 1.4.0. With 1.4.0 this page was also planned to be deleted. Therefore there is no point in moving it to a separate page with this change.

* Remove zeebe-process-test warning

By the time this page will be released (1.4.0) there will be a stable version

* Rewrote Zeebe Process Test documentation

A lot has been changed since I initially wrote this documentation. This change updates it to the definitive 1.4 version.
I have also renamed the page to reflect the title

* Update zeebe-process-test.md

technical documentation review

* Fix broken url

* Add zeebe-process-test redirects

The java-client/testing page was renamed to java-client/zeebe-process-test. Redirects need to be added to ensure the java-client/testing page navigates to the new page.

Co-authored-by: Christina Ausley <christina.ausley@camunda.com>
Co-authored-by: christinaausley <84338309+christinaausley@users.noreply.github.com>
Co-authored-by: Amara <amara.graham@camunda.com>