docs(grpc): remove yaml deprecation note
This note existed to inform users of the DeployProcess RPC that the YAML
format was deprecated and expected to be removed at some point in the
future. This removal has already happened with the release of version
1.0. At that time, this note could've been removed already.

As the DeployProcess RPC has now been completely deprecated, and YAMLs
haven't been supported for more than a year, it makes sense to remove
this note. It is no longer needed.