fix(announcements): re-add removed header
PR camunda-cloud/camunda-cloud-documentation#338 made some changes to
the announcements. It looks like an effort was made to re-order the
headers chronologically. However, in doing so, one of the deprecation
headers was removed and a line of text was duplicated.

This rectifies these changes, by re-adding the header and re-orders the
section headers according to the above assumption (chronological order).
It also removes the duplicated line.