docs: use better job type
Since DMN decision is the default implementation, job workers are
unlikely to be used to evaluate dmn decisions. The example should
therefore reflect an example scenario that uses a job worker. The
example xml already used the concept of calculating risk. We can simply
improve that example by clarifying the job type.