Document business rule task with called decision (#661)
* docs(bpmn): describe business rule task with called decision

Camunda Cloud 1.4 will add support for business rule tasks with a called
decision extension element.

Existing business rule tasks (using the taskDefinition) still work as
expected, but are referred to as alternative task implementation. This
is to help users understand the difference when chosing the
implementation in the Modeler. All that field does is switch between the
`taskDefinition` and `calledDecision` extension elements.

The modeler will make the `calledDecision` the default implementation.
So it makes sense that the docs takes a similar approach. I've therefore
tried to write from the perspective of using the business rule task to
evaluate decisions modeled as DMN decision. And I've provided an info
box near the top to indicate that an alternative implementation exists
as well.

Additional effort was made to the order of execution between input
mappings, decisionId expression evaluation and decision evaluation, as
well as what happens when an incident at the business rule task is
resolved.

At this time, I've chosen not to document any restrictions we impose on
the decisionId (like spaces). We should consider whether documenting
this is necessary.

I've also chosen not to directly reference the DMN scala engine as the
internal decision engine. I think this should be independent from the
business rule task. However, the choise of decision engine does impose
some limitations as users migrating from Camunda Platform 7 will
discover. We should consider where these limitations should be
documented.

* docs(concepts): add decision to incident cases

A decision evaluation can also lead to an incident. Not all cases are
documented, but decisions are a major part of Camunda Cloud from 1.4
onwards, and so it makes sense to document this here.

* docs: favor camunda platform 8 over camunda cloud

* docs: favor job worker impl over alt task impl

It makes more sense for the section header to reflect the naming used in
the modeler, which talks about a job worker implementation and a dmn
decision implementation.

* docs: use better job type

Since DMN decision is the default implementation, job workers are
unlikely to be used to evaluate dmn decisions. The example should
therefore reflect an example scenario that uses a job worker. The
example xml already used the concept of calculating risk. We can simply
improve that example by clarifying the job type.

* docs: cleanup notes

Not all notes are needed anymore, and the ones that are useful should be
made more prominent by using info over note.

* docs: favor process instance over execution

* docs: favor link over bolding

* style(formatting): technical review

technical review

Co-authored-by: christinaausley <84338309+christinaausley@users.noreply.github.com>