Document new DeployResource RPC and its usage (#716)
* docs(grpc): add DeployResource rpc

The DeployResource RPC has been recently introduced and will be
available from 8 onwards. It can be used to deploy any resource (e.g.
BPMN process files and DMN decision models). In the future, we may
support additional resources.

This RPC replaces the existing DeployProcess RPC, which is deprecated
since 8.

* docs(grpc): deprecate DeployProcess rpc

The DeployProcess RPC has been deprecated since 8. Users should switch
to the more flexible DeployResource RPC. Of course the documentation
should describe this.

To fit it in this page, a new Deprecated RPCs heading is added. The
deprecated RPC has been moved to it and a deprecation note has been
added explaining since when it was deprecated and that is has been
replaced by the DeployResource RPC.

* docs(grpc): remove yaml deprecation note

This note existed to inform users of the DeployProcess RPC that the YAML
format was deprecated and expected to be removed at some point in the
future. This removal has already happened with the release of version
1.0. At that time, this note could've been removed already.

As the DeployProcess RPC has now been completely deprecated, and YAMLs
haven't been supported for more than a year, it makes sense to remove
this note. It is no longer needed.

* fix(announcements): re-add removed header

PR camunda-cloud/camunda-cloud-documentation#338 made some changes to
the announcements. It looks like an effort was made to re-order the
headers chronologically. However, in doing so, one of the deprecation
headers was removed and a line of text was duplicated.

This rectifies these changes, by re-adding the header and re-orders the
section headers according to the above assumption (chronological order).
It also removes the duplicated line.

* docs(announcements): deprecate deployprocess rpc

The DeployProcess RPC is deprecated since 8. This is documented in the
gRPC API documentation, but should also be announced in the
announcements.

* docs: use zbctl deploy resource

Since `zbctl deploy <process>` has been deprecated for removal. Any
usage should be replaced by the new `zbctl deploy resource <resource>`

* docs(go-client): use newDeployResourceCommand

The `newDeployResourceCommand` is the new way to deploy resources. The
old way is deprecated for removal and we should not use it in
documentation.

* docs(java-client): use newDeployResourceCommand

Switch from the newDeployCommand to newDeployResourceCommand, as one
replaces the other. The newDeployCommand is deprecated for removal and
should no longer be used in documentation.

* docs(go-client): update getting started to v8 client

Updates the getting started go-client guide to use the new v8 go client.

It also switches to using the new DeployResource RPC. For this the
outputs have been updated by running the steps against a 8.0.0 cluster.

* docs(go-client): update quick ref to v8 client

Similar to the previous commit, this updates the quick reference to use
the new v8 go-client.

* docs(grpc): move error handling above deprecated rpcs