build: add prettier
Prettier allows us to automatically format our code, config,
documentation, and other text files.

Prettier applies an opinioned set of rules which can be modified in the
.prettierrc.json file.

To make sure that we don't have an enormous list of changes to review in
one go, I've added all files in this repo to the .prettierignore file.
Files that we want to format can be removed from this list. However, in
some cases we want to never format specific files. The node_modules
folder is a good example of this. Therefore, I've placed it in the top
of the prettierignore file. We should carefully consider each file
whether we want to format it or not. I expect that at some point the
list of files under the block comment is empty and these files have
either been moved to the top part of the file (next to node_modules) or
have been removed altogether from the prettierignore file.

Note that you can use filenames, dirnames and glob patterns to describe
the files that should be ignored.

See: https://prettier.io/docs
See: https://prettier.io/docs/en/option-philosophy.html
See: https://prettier.io/docs/en/options.html