style: format npm files
The npm files (package files and prettier config files) should be
correctly formatted as well.

An exception should be made to the package.lock file as it is generated
by npm and we don't want to mess with it.

Note that the other files were all already correctly formatted, but from
now on they will be correctly formatted.