style: format github files
Formats the github files, incidently only changing the workflow files.

Here's the first case where I wondered whether the change is correct.
Note the change in .github/workflows/publish-stage.yaml on line 23. It
replaces the double quotes with single ones to remove some of the
escaping that was necessary. However, I wondered if the 2 consequtive
single quotes work correctly. Perhaps this is fine, but its also hard to
test this line change.

I searched for this case and discovered that this is part of the YAML
spec: single quotes are escaped using another single quote. So it's all
good, but we should always be careful about the automatically applied
formatting rules.

See https://yaml.org/spec/1.2.2/#732-single-quoted-style