fix: remove unused blog feature
Next on the list of files to format with prettier was the blog. This is
a directory that contains files that should be part of the docusaurus
blog feature. However, this feature is not used by the cc-docs. It is
disabled in the docusaurus.config.js file. Lets remove this clutter.