style: format media-src folder
It's not entirely clear for me if we will have files in this folder that
we want to format. For now, it only contains bpmn and pptx files which
are both not formatted by prettier. If we happen to add files to this
folder that shouldn't be formatted then we can always ignore those files
again.