style: format static folder
Now we get to the first real javascript code formatting. These files
should be formatted, because we maintain them and make changes to them
once in a while.

Note that some of the formatting rules that are applied feel a bit
silly, like calls to console.log expanded over 4 lines because of
excessive line wrapping. We should consider whether we want to change
this.