Enable prettier to format files (#815)
* build: add prettier

Prettier allows us to automatically format our code, config,
documentation, and other text files.

Prettier applies an opinioned set of rules which can be modified in the
.prettierrc.json file.

To make sure that we don't have an enormous list of changes to review in
one go, I've added all files in this repo to the .prettierignore file.
Files that we want to format can be removed from this list. However, in
some cases we want to never format specific files. The node_modules
folder is a good example of this. Therefore, I've placed it in the top
of the prettierignore file. We should carefully consider each file
whether we want to format it or not. I expect that at some point the
list of files under the block comment is empty and these files have
either been moved to the top part of the file (next to node_modules) or
have been removed altogether from the prettierignore file.

Note that you can use filenames, dirnames and glob patterns to describe
the files that should be ignored.

See: https://prettier.io/docs
See: https://prettier.io/docs/en/option-philosophy.html
See: https://prettier.io/docs/en/options.html

* style: format babel.config.js

This is the first file we've automatically formatted with prettier. By
removing the line from .prettierignore, prettier is able to apply its
formatting rules on the file. I'll try to make these changes in small
increments to ease reviewing.

* style: format repo docs

Formats the repository documentation files like the readme and
contributing guide. Note that most files were already correctly
formatted according to prettier, with an exception to the
code-of-conduct.md file. This may happen more often when we start
formatting more and more files.

* style: format docusaurus config files

This formats the docusaurus config files using prettier, by removing the
entries from the .prettierignore file. Slowly, we can apply these rules
on more and more files. Note that the applied rules are defined by
prettier. If we feel that certain rules don't make sense for us, we can
change it using the options in the .prettierrc file.

See: https://prettier.io/docs/en/option-philosophy.html
See: https://prettier.io/docs/en/options.html

* fix(sidebars): re-add removed lines

These lines were accidentally removed while resolving conflicts.

* style: format npm files

The npm files (package files and prettier config files) should be
correctly formatted as well.

An exception should be made to the package.lock file as it is generated
by npm and we don't want to mess with it.

Note that the other files were all already correctly formatted, but from
now on they will be correctly formatted.

* style: format git files

Mostly we want to format the .gitignore file. The files in .git/ folder
are managed by git itself and are generally not touched by human hands.

* style: ignore vscode files

We don't want to format the files that are managed by vscode. It's
perfectly capable to decide what those files should look like.

* style: format github files

Formats the github files, incidently only changing the workflow files.

Here's the first case where I wondered whether the change is correct.
Note the change in .github/workflows/publish-stage.yaml on line 23. It
replaces the double quotes with single ones to remove some of the
escaping that was necessary. However, I wondered if the 2 consequtive
single quotes work correctly. Perhaps this is fine, but its also hard to
test this line change.

I searched for this case and discovered that this is part of the YAML
spec: single quotes are escaped using another single quote. So it's all
good, but we should always be careful about the automatically applied
formatting rules.

See https://yaml.org/spec/1.2.2/#732-single-quoted-style

* fix: remove unused blog feature

Next on the list of files to format with prettier was the blog. This is
a directory that contains files that should be part of the docusaurus
blog feature. However, this feature is not used by the cc-docs. It is
disabled in the docusaurus.config.js file. Lets remove this clutter.

* style: format hacks

Doesn't really lead to any changes, but when we add new files to the
hacks folder, we probably want to format them.

* style: format media-src folder

It's not entirely clear for me if we will have files in this folder that
we want to format. For now, it only contains bpmn and pptx files which
are both not formatted by prettier. If we happen to add files to this
folder that shouldn't be formatted then we can always ignore those files
again.

* style: no need to ignore txt files

These are not formatted anyways by prettier. We can simply remove the
ignore entry.

* style: format static folder

Now we get to the first real javascript code formatting. These files
should be formatted, because we maintain them and make changes to them
once in a while.

Note that some of the formatting rules that are applied feel a bit
silly, like calls to console.log expanded over 4 lines because of
excessive line wrapping. We should consider whether we want to change
this.