docs(concepts): describe start instance anywhere
C8.1 supports a new feature which allows users to create process
instances starting at user-defined elements instead of at the default
none start event.

I've tried to describe how the feature can be used, why it exists and
that it should be used with care.

I've explicitly not described how to use it as a workaround for the
missing Process Instance Migration feature. At the time of writing I
think that this does not belong in the docs. Instead, we can describe
this workaround in the feature request issue.

I wondered whether the start instructions should link to the gRPC
definition of the CreateProcessInstance RPC to show what a start
instruction looks like, but I'm not sure about it.