Document create process instance at user-defined elements (#1023)
* docs(concepts): clean up instance creation intro

This intro was using some strange language like 'At Camunda', while
these concepts are specific to Camunda Platform 8. And message event
handler is likely a legacy concept that no longer exists. In C8 messages
are not handled, but published, correlated and deleted.

In addition, the intro can function as a jumping off point (faster
navigation) using links.

* docs(concepts): describe start instance anywhere

C8.1 supports a new feature which allows users to create process
instances starting at user-defined elements instead of at the default
none start event.

I've tried to describe how the feature can be used, why it exists and
that it should be used with care.

I've explicitly not described how to use it as a workaround for the
missing Process Instance Migration feature. At the time of writing I
think that this does not belong in the docs. Instead, we can describe
this workaround in the feature request issue.

I wondered whether the start instructions should link to the gRPC
definition of the CreateProcessInstance RPC to show what a start
instruction looks like, but I'm not sure about it.

* docs(concepts): add code example to start anywhere

This might help users to understand the start before element instruction
better. However, the Java Client's API differs from the gRPC API.

* style(formatting): technical review

* style(formatting): fix typo

Co-authored-by: Amara <amara.graham@camunda.com>
Co-authored-by: Christina Ausley <christina.ausley@camunda.com>