Follow-up on inclusive gateway documentation (#1247)
* docs(bpmn): or gateway decisions are based on data

Make it clear that decisions are based on data (process variables). This
aligns the documentation with that of the XOR gateway.

* docs(bpmn): improve grammar of OR gateway support notice

In the notice, we shouldn't refer to 'we' but rather Camunda Platform 8.
In addition, it makes sense to explain that the AND/XOR gateway can be
used to mimic the missing OR merging behavior.

* docs(bpmn): improve grammar on OR gateway default flow

This makes it a bit more clear that the default flow is not mandatory,
but optional.

* docs(bpmn): refer to condition instead of expression

The conditionExpression is introduced later in the text (in the
condition paragraph). So for now, we can simplify it by just referring
to condition.

In addition, all conditions of the outgoing sequence flows are
evaluated. This wasn't very clear before.

* docs(bpmn): improve the or gateway alt-texts

The alt texts could use some work to describe the essence of what they
intent to show.

* docs(bpmn): combine OR gateway flow requirements

There are some restrictions on the outgoing flows of the inclusive
gateway. For modelers it makes sense to see these rules as one package.

* docs(bpmn): clarify no default flow behavior

It should be clear that an incident is only created when no condition is
fulfilled.