feat: support sharing media pre-sources
Many diagrams on these docs are screenshots taken from BPMN files.
When we want to make changes to these, we needed to recreate these BPMN
files from the data shown on the screenshot.

This new folder offers a place to share these files without them being
hosted by the Docusaurus server.

Note, that in the future we want to automate taking screenshots. This
folder with pre-sources can be used as the input for taking those
automated screenshots.

While the screenshots are still taken manually, we need to make sure
that these BPMN files are kept up-to-date when making changes to the
diagrams themselves. This addition therefore pushes for a faster
adoption of the automated screenshotting.