fix(docs): remove default from main or gateway example
The main example for the Inclusive (OR) Gateway should not have a
default marker to start simple. The default flow is introduced later in
the document.

The examples showing the behavior are screenshots taken from Operate.

Note that this commit also adds the pre-source of the image to the
`media-pre-src` folder. This makes it easier to make changes to the
image without having to remodel the process in Camunda Modeler.