fix(docs): dont eval the default flows condition
The default flow is taken when none of the conditions are fulfilled. It
cannot itself have a condition.

Actually it can have a condition, but we discourage it and Zeebe doesn't
evalute that condition.It must be clear to the user that the condition
is not evaluated.

This commit also removes the label from the default branch to highlight
that there is no condition on this sequence flow.