Document inclusive gateway behavior changes (#1269)
* feat: support sharing media pre-sources

Many diagrams on these docs are screenshots taken from BPMN files.
When we want to make changes to these, we needed to recreate these BPMN
files from the data shown on the screenshot.

This new folder offers a place to share these files without them being
hosted by the Docusaurus server.

Note, that in the future we want to automate taking screenshots. This
folder with pre-sources can be used as the input for taking those
automated screenshots.

While the screenshots are still taken manually, we need to make sure
that these BPMN files are kept up-to-date when making changes to the
diagrams themselves. This addition therefore pushes for a faster
adoption of the automated screenshotting.

* fix(docs): remove default from main or gateway example

The main example for the Inclusive (OR) Gateway should not have a
default marker to start simple. The default flow is introduced later in
the document.

The examples showing the behavior are screenshots taken from Operate.

Note that this commit also adds the pre-source of the image to the
`media-pre-src` folder. This makes it easier to make changes to the
image without having to remodel the process in Camunda Modeler.

* fix(docs): use list contains(/2)

The built-in FEEL function is called `list contains` not `contains`.

* fix(docs): dont eval the default flows condition

The default flow is taken when none of the conditions are fulfilled. It
cannot itself have a condition.

Actually it can have a condition, but we discourage it and Zeebe doesn't
evalute that condition.It must be clear to the user that the condition
is not evaluated.

This commit also removes the label from the default branch to highlight
that there is no condition on this sequence flow.