feat(reference): Add definition of early access releases
The docs are lacking cross-product definitions for early access releases
. While multiple products mention "preview", "alpha" and "developer
preview" releases, all those mentions define them in different ways.
This is an attempt to unify and align these definitions on a
cross-product scale, to make it more user-friendly, unambiguous and also
to act as a point of reference for the upcoming Web Modeler Self-Managed
developer preview.