Remove reserved "return" keyword from CLI getting started guide (#1310)
* fix(apis): remove reserved "return" keyword from guide

The example in this guide makes use of the reserved "return" keyword as a variable. This causes the process to be undeployable, and thus is a bad example.
Because of this "return" has been changed to "result".

All the images in the guide have been updated to reflect this change and to be more up-to-date with recent versions of the modeller and operate.

* fix(apis): remove reserved "return" keyword from guide version 8.0

The example in this guide makes use of the reserved "return" keyword as a variable. This causes the process to be undeployable, and thus is a bad example.
Because of this "return" has been changed to "result".

All the images in the guide have been updated to reflect this change and to be more up-to-date with recent versions of the modeller and operate.

* fix(apis): remove reserved "return" keyword from guide version 1.3

The example in this guide makes use of the reserved "return" keyword as a variable. This causes the process to be undeployable, and thus is a bad example.
Because of this "return" has been changed to "result".

All the images in the guide have been updated to reflect this change and to be more up-to-date with recent versions of the modeller and operate.

Co-authored-by: Amara <amara.graham@camunda.com>