docs(feel): change the language of the code blocks to `feel`
Change the language of the code blocks from `js` to `feel` to avoid that the FEEL expressions are formatted by Prettier. The formatting would make the expressions invalid.

The downside is that we don't have a syntax highlighting of the FEEL expressions anymore. But we can work on this in the future.

Alternatively, we would have to add ignore blocks for Prettier around each code block. That would be a lot of boiler code and easy to miss for new code blocks.