docs(feel): Update the FEEL docs to 1.15 (#1323)
* docs(feel): copy docs from FEEL-Scala repository of version 1.15

Update the FEEL docs with the docs from the FEEL-Scala repository of version 1.15.

Fix code blocks that were formatted by Prettier.

* docs(feel): change the language of the code blocks to `feel`

Change the language of the code blocks from `js` to `feel` to avoid that the FEEL expressions are formatted by Prettier. The formatting would make the expressions invalid.

The downside is that we don't have a syntax highlighting of the FEEL expressions anymore. But we can work on this in the future.

Alternatively, we would have to add ignore blocks for Prettier around each code block. That would be a lot of boiler code and easy to miss for new code blocks.

* docs(feel): change the language of the code blocks to `feel`

Adopt the changes from the current/next version:
- Change the language of the code blocks from `js` to `feel` to avoid that the FEEL expressions are formatted by Prettier
- Fix the code blocks that were formatted