Updates correlationKey in sample xml (#1306)
* Updates correlationKey in sample xml

I believe in the descriptions of this example above, you reference the variable orderId as the place that holds the correlationKey. However in this xml example, it appears that it is referencing a correlation key of the static value "orderId". I have updated the sample xml to reflect the reference to a variable instead. If I'm wrong, then I have a few question I need answered.

* Adds orderId variable as the correlationKey

* Adds orderId variable as the correlationKey