Update slack.md
Had some special cases that blocked my slack configuration for the slack connector during camunda hackday.
First I had to create a workspace in slack to be allowed to install an app to it. A guest status would not be enough.
After all I also need to invite the slack app in the workspace.
Hope the additional hints help other people to faster configure a slack app.