docs(zeebe): specify private key format
Adds a note to specify that private keys for Zeebe (both broker and
gateway) must be in the PKCS8 format, and explain how to convert
existing private keys to this format.

For the record, this is a Netty limitation, and Netty is used both by
the intra-cluster communication layer and grpc-java (for the clients).
See https://netty.io/wiki/sslcontextbuilder-and-private-key.html