style: use orangemunda as primary color (#1698)
This switches the Infima primary colors to shades of orange-munda.
The shades are produced using Docusaurus' own tool.

It is unclear to me why #ff1f00 was used as the primary color before.
It might be related to contrast to improve the accessibility. However,
it may also simply be a leftover from the start of the
camunda-platform-docs.