Clarify acronyms and defaults on backpressure (#1721)
* Clarify acronyms

I'm trying to understand how back pressure can be configured. First hurdle are the unknown acronyms RTT and AIMD.

* Add link to default values and clarify AIMD 

Add a link to the default values points to the Zeebe repo. I think this better than copying the variables and their values.

From the configuration template, I could get the missing words to describe the AIMD algorithm.

* Improve language

* Add cluster configuration

Not only the standalone broker. Some values differ in the backpressure section, like fixed limit.

* Adjust section level of Gradient and Gradient2

Both are algorithms equally to the other three.

* style(formatting): technical review

---------

Co-authored-by: Christina Ausley <christina.ausley@camunda.com>