docs: describe self-signed certificates workaround
While the Elasticsearch exporter does not yet support self-signed
certificates explicitly, we want to document the workaround described by
@vctrmn in https://github.com/camunda/camunda-platform-helm/issues/473

This may help users that have to connect using a self-signed
certificate, for example those using Elastic Cloud.

The note describes what must be done, and provides an example for those
using k8s to deploy their Zeebe instances.

I've chosen to add this documentation to 8.2 (next), 8.1, and 8.0, as I
believe it is relevant in all those cases.