docs(bpmn): describe link events
This adds a description of link events using multiple examples.

Note, that it is uncommon in the BPMN docs to use several examples, but
it makes sense as we're shifting these docs from reference to a focus on
users and use cases.

Both examples are added as media-pre-src, so they are available to make
changes and reproduce the screenshots.

Note, that in C8 link events DO NOT YET support linking between
different "pages" (i.e. processes).

See:
- https://camunda.com/bpmn/reference/#events-link