Document Opensearch and Elasticsearch exporters (#1837)
* docs(exporters): add Opensearch exporter page

Copies the configuration instructions from the Zeebe README to it's dedicated page in the official documentation.

This exporter is officially supported and thus it makes sense to include it in the official documentation. The README in Zeebe will be changed to refer to this page.

* docs(exporters): add Elasticsearch exporter page

Copies the configuration instructions from the Zeebe README to it's dedicated page in the official documentation.

This exporter is officially supported and thus it makes sense to include it in the official documentation. The README in Zeebe will be changed to refer to this page.

* docs(operate): modify link to elasticsearch exporter from Zeebe README to docs

The documentation of this exporter has been moved into the official docs. We should prefer this over linking to the Zeebe README

* docs(supported-envs): remove message about not having Opensearch support

* style(formatting): technical review

* docs(es-exporter): add retention policy configuration

Adds the configuration options for configuring a retention policy. This is available as of Zeebe 8.2.

---------

Co-authored-by: Christina Ausley <christina.ausley@camunda.com>