docs(bpmn): explain how error variables are merged into the process
Error variables are merged differently compared to the bahavior of most
other elements. Typically, variables are merged into the process
instance and variable mappings can be used to change this bahavior.
However, for error catch events, the variables are merged into the catch
event itself, and variables mappings can be used to merge them further
into the process instance. This should be clarified by the docs so users
understand what to expect.

The reason of this difference is that the variables of an error
typically describe details only relevant to that specific part of the
process, not to the entire process instance. By localizing the
variables, they don't overwrite any existing variables that may have
lead to the error, while output mappings still provide control to the
user to decide merging them into the process instance for further usage.