Document throwing errors with variables (#1906)
* docs(apis): copy throw error rpc description

Directly copy the current ThrowError RPC description defined by the
`gateway.proto` file (the definition of the API).

This description provides a wording that is more focused to usage than
the previous one.

* docs(apis): copy new gRPC ThrowErrorRequest

Variables can now be passed along with the thrown error. The details are
copied directly from the `gateway.proto` gRPC API definition.

* docs(apis): explain what error variables are for

Users should understand what the error variables can be used for.

* docs(bpmn): explain how error variables are merged into the process

Error variables are merged differently compared to the bahavior of most
other elements. Typically, variables are merged into the process
instance and variable mappings can be used to change this bahavior.
However, for error catch events, the variables are merged into the catch
event itself, and variables mappings can be used to merge them further
into the process instance. This should be clarified by the docs so users
understand what to expect.

The reason of this difference is that the variables of an error
typically describe details only relevant to that specific part of the
process, not to the entire process instance. By localizing the
variables, they don't overwrite any existing variables that may have
lead to the error, while output mappings still provide control to the
user to decide merging them into the process instance for further usage.

* style(formatting): technical review

* docs(bpmn): adjust wrong link to variable mappings

This seemed to have been a wrongful copy-paste. We want to link to
variable mappings so users can learn about that concept if they haven't
already.

---------

Co-authored-by: Christina Ausley <christina.ausley@camunda.com>
Co-authored-by: christinaausley <84338309+christinaausley@users.noreply.github.com>