feat: Update FEEL docs (#1945)
* feat(feel): Update FEEL docs for NEXT

Copy the latest FEEL docs (version 1.16) from the feel-scala project.

The latest version includes a new format to document the built-in functions (split into description, function signature, and examples), and markers for functions/features that are not part of the DMN specification but an extension from Camunda.

Copy the FEEL docs from the NEXT version to version 8.2. Both versions depend on the same version of the FEEL engine.

Add the Camunda extension markers in the FEEL docs for versions 8.0 and 8.1. These markers are indicators that a function or feature is not part of the DMN specification but an extension of Camunda.