docs(zeebe): update grafana dashboard
This updates the screenshot of the grafana dashboard and links to a
publicly available read-only snapshot for users to play around with.

I've chosen to use the same screenshot and snapshot for 8.0, 8.1, 8.2
and next. While there are some differences between dashboard versions,
we expect users to always use the latest version anyway.

I haven't updated older version since these are no longer officially
maintained and old enough that metrics and dashboards might have
changed significantly.