docs(timers): document support for time date events
As of 8.3 Zeebe officially supports Timers defined with a time date in intermediate and boundary events.

For non-interrupting boundary events there is a slight caveat. Usually these are repeatable and can be triggered multiple times. For a time date the event will only be triggered once, as this is not repeatable.