Update the FEEL documentation for version 8.3 (#2658)
* docs: Update FEEL documentation

Camunda 8.3 uses a new version of FEEL-Scala. Copy the FEEL documentation from FEEL-Scala to add the updated content.

* docs: Extend the update guide to 8.3

Mention the new error handling of expressions and the new semantics of the function `is defined()` in the update guide.

* docs: Update incident-related content

With the new error handling of expressions, incidents are not created in all cases anymore. For example, variable mappings doesn't cause incidents but evaluate missing variables to null.

* docs: Update expressions page

Instead of describing a few FEEL expressions and their usage, link to the FEEL documentation. This removes duplicated content and increases the maintainability.

* docs: Fix typo

Co-authored-by: Nicola Puppa <72783243+nicpuppa@users.noreply.github.com>

* docs: Update expressions page

Currently, the text mentions a few attributes but not all of them. There is no value in having a complete list of attributes here.

Removing the list of attributes to increase the maintenance.

---------

Co-authored-by: Nicola Puppa <72783243+nicpuppa@users.noreply.github.com>