Document signal catch and throw events (#2708)
* docs(concepts): explain how to broadcast signals

There are now two ways to broadcast signals. In 8.2, only the RPC could
be used, but in 8.3 we've introduced signal throw events and signal end
events.

These events don't have their own documentation paragraphs yet, so I'll
link them later.

I've taken the opportunity to rewrite some sections to reuse specific
words like broadcast and trigger.

* docs(concepts): caution signal performance impact

Signal can have a negative performance impact starting from C8.3, where
signals are broadcasted to all partitions and can trigger signal
subscriptions of process instances.

We must teach users about this impact. While I'd love to see some best
practices for signal events in the future, I think it's enough for now
to document the performance impact as reference details.

* docs(bpmn): process with signal start event deleted

In 8.3, we introduce the ability to delete resources
(processes/decisions). This can have an affect on signal start events
when the process that contains them is deleted, or when the previous
version contains them.

* docs(bpmn): describe signal catch/throw events

With 8.3, new signal catch and throw events have been introduced. I've
added a description for them, and marked them as covered.

* docs(bpmn): add signals to event-based-gateway

Intermediate signal events can also be attached to event-based-gateways.
I've updated the description, example, and image.

* style(formatting): technical review

* fix: remove "recommendation" from the warning

It's hard to tell users when to use signals and when they shouldn't. It's very dependent on independent use-cases. Instead of recommending, we should aim to inform the user. This is what this warning already does. With this information users can decide themselves if this fits their use-case.

---------

Co-authored-by: christinaausley <84338309+christinaausley@users.noreply.github.com>
Co-authored-by: Christina Ausley <christina.ausley@camunda.com>
Co-authored-by: Remco Westerhoud <remco@westerhoud.nl>