Document multi-tenancy for Zeebe gRPC (#2696)
* feat(apis-tools): document tenant-related grpc properties

Document the tenantId and tenantIds properties in all of the
gRPC request and response messages that were made tenant-aware.

In some gRPC requests the tenantId property is not required as
the tenant ownership of the data is already clear within Zeebe.

* feat(apis-tools): document tenant-related grpc errors

Document the multi-tenancy-related errors that may be produced
by tenant-aware gRPC requests.

* doc(api-tools): document multi-tenancy limiations of zeebe grpc spec

Document the limitations of multi-tenancy in the Zeebe gRPC
calls. Signals are not supported for non-<default> tenants,
and resource deletion is not tenant-aware, meaning that it
can delete any resource no matter the tenant ownership.

* doc(api-tools): multi-tenancy for deprecated rpc

Document that for the deprecated DeployProcess RPC, when
multi-tenancy is enabled, any deployments are owned by
the <default> tenant. Deploying to other tenants through
this RPC is not possible since it's deprecated and no
new behaviors are added.

* fix(grpc): correct multi-tenancy support note for deletion

Correct the note on multi-tenancy support for resource deletion in
the gRPC documentation. Resource deletion is only supported for the
<default> tenant.